pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mjcloud"
version = "0.1.0"
description = "One-click MuJoCo GPU simulation environments in the cloud"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "click>=8.0",
    "google-cloud-compute>=1.14",
    "google-auth>=2.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
mjcloud = "mjcloud.cli:cli"

[project.optional-dependencies]
dev = ["pytest", "ruff"]
